preflight-0.3.1
This libary was extended by rafael zuim (this git owner) and Evandro silva. The project is origal started and develped by YOB https://github.com/yob/ , but we try contact and have no response by email. We want to contribute with his lib. BTW this lib is so usefull and we rebuild it and use in us context.